// src/service/file/schemas/lectureReportAddRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { LectureReportAddRequestCourse } from './lectureReportAddRequestCourse';
import type { LectureReportCardForAdd } from './lectureReportCardForAdd';

export interface LectureReportAddRequest {
  /**
   * 놀이보고서의 주제.
   * @maxLength 20
   */
  subject: string;
  /** 놀이활동의 대상 연령 */
  studentAge: number;
  /** 놀이활동의 대상 과정 */
  course?: LectureReportAddRequestCourse;
  /**
   * 놀이활동 시작일(yyyy-MM-dd)
   * @maxLength 10
   */
  startsAt: string;
  /**
   * 놀이활동 종료일(yyyy-MM-dd)
   * @maxLength 10
   */
  endsAt: string;
  /**
   * 놀이 속 배움에 들어갈 내용
   * @maxLength 200
   */
  learningPoint: string;
  /**
   * 교사지원에 들어갈 내용
   * @maxLength 200
   */
  teacherSupport: string;
  /**
   * 소유자 계정 아이디
   * @minimum 0
   */
  ownerAccountId: number;
  /**
   * 소유자 프로필 아이디
   * @minimum 0
   */
  ownerProfileId: number;
  /**
   * 최초 생성자 계정 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorAccountId: number;
  /**
   * 최초 생성자 프로필 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorProfileId: number;
  /**
   * 생성자의 ip
   * @minLength 1
   * @maxLength 40
   */
  creatorIp: string;
  /** 하위 카드 리스트 */
  lectureReportCards?: LectureReportCardForAdd[];
}