// src/service/file/schemas/lectureReportResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { LectureReportResultCourse } from './lectureReportResultCourse';
import type { LectureReportCardResult } from './lectureReportCardResult';
import type { DriveItemResult } from './driveItemResult';
import type { SmartFolderItemResult } from './smartFolderItemResult';

/**
 * 놀이보고서면 놀이보고서 정보
 * @nullable
 */
export type LectureReportResult = {
  /** 놀이보고서의 아이디 */
  id: number;
  /**
   * 놀이보고서의 주제.
   * @maxLength 20
   */
  subject: string;
  /** 놀이활동의 대상 연령 */
  studentAge: number;
  /** 놀이활동의 대상 과정 */
  course: LectureReportResultCourse;
  /**
   * 놀이활동 시작일(yyyy-MM-dd)
   * @maxLength 10
   */
  startsAt: string;
  /**
   * 놀이활동 종료일(yyyy-MM-dd)
   * @maxLength 10
   */
  endsAt: string;
  /** 놀이 속 배움에 들어갈 내용 */
  learningPoint: string;
  /** 교사지원에 들어갈 내용 */
  teacherSupport: string;
  /**
   * 소유자 계정 아이디
   * @minimum 0
   */
  ownerAccountId: number;
  /**
   * 소유자 프로필 아이디
   * @minimum 0
   */
  ownerProfileId: number;
  /**
   * 최초 생성자 계정 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorAccountId: number;
  /**
   * 최초 생성자 프로필 아이디(공유된 것을 저장하는 게 아니라면 소유자와 동일, 공유된 것이라면 공유한 사람의 정보)
   * @minimum 0
   */
  creatorProfileId: number;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt: string;
  /** 하위 카드 리스트 */
  lectureReportCards?: LectureReportCardResult[];
  driveItem?: DriveItemResult;
  smartFolderItem?: SmartFolderItemResult;
} | null;