// src/service/file/schemas/smartFolderItemResultFileType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 자료의 종류. URL,DOCUMENT,IMAGE,VIDEO,AUDIO,ARCHIVE,LECTURE_PLAN, ... ,ETC
 */
export type SmartFolderItemResultFileType =
  (typeof SmartFolderItemResultFileType)[keyof typeof SmartFolderItemResultFileType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SmartFolderItemResultFileType = {
  FOLDER: 'FOLDER',
  URL: 'URL',
  DOCUMENT: 'DOCUMENT',
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
  AUDIO: 'AUDIO',
  ARCHIVE: 'ARCHIVE',
  LECTURE_PLAN: 'LECTURE_PLAN',
  LECTURE_PLAN_REPORT: 'LECTURE_PLAN_REPORT',
  STUDENT_RECORD: 'STUDENT_RECORD',
  STORY_BOARD: 'STORY_BOARD',
  STUDENT_CLASSIFICATION: 'STUDENT_CLASSIFICATION',
  ACTIVITY_CLASSIFICATION: 'ACTIVITY_CLASSIFICATION',
  PHOTO_COMPOSITION: 'PHOTO_COMPOSITION',
  PRIVATE_DATA_ENCRYPTION: 'PRIVATE_DATA_ENCRYPTION',
  SKETCH_CREATION: 'SKETCH_CREATION',
  PHOTO_ALBUM: 'PHOTO_ALBUM',
  AI_WRITING: 'AI_WRITING',
  STUDENT_AND_ACTIVITY_CLASSIFICATION: 'STUDENT_AND_ACTIVITY_CLASSIFICATION',
  TEXT_MEMO: 'TEXT_MEMO',
  ETC: 'ETC',
} as const;